Feature/qr code (#6)
* Add migration of table (add new index - Link)

* QRCodeHelper

* FrontEnd

* Add QRCode realization on TableController

* Fix bug with QRCode

* fix repositories bugs

* change to dishDto and tableDto models and create crud

+++ Restraunt.WebAPI/Controllers/TableController.cs
@@ -3,8 +3,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using QRCoder;
 using Restraunt.Core;
+using Restraunt.Core.Dto;
 using Restraunt.Data;
 using Restraunt.Data.Interfaces;
+using Restraunt.Data.Repositories;
 using System.Data;
 using System.Drawing;
 
@@ -29,8 +31,6 @@ namespace Restraunt.WebAPI.Controllers
         [HttpGet]
         public async Task<ActionResult<List<Table>>> Get()
         {
-           
-
             return Ok(await _tableRepository.Select());
 
         } 
@@ -39,9 +39,9 @@ namespace Restraunt.WebAPI.Controllers
         [HttpGet("{Id}")]
         public async Task<ActionResult<Table>> Get(Guid Id)
         {
-           
-
-           
+            var entity = await _tableRepository.Get(Id);
+            if (entity == null)
+                return BadRequest("Entity not found");
 
             return Ok(await _tableRepository.Get(Id));
 
@@ -50,13 +50,15 @@ namespace Restraunt.WebAPI.Controllers
 
         
          [HttpPost]
-        public async Task<ActionResult<List<Table>>> AddTable(Table table)
+        public async Task<ActionResult<List<Table>>> AddTable(TableDto table)
         {
             if (ModelState.IsValid)
             {
                 
 
-           table.Link = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host}/Table/{table.Id.ToString()}";
+
+           table.Link = $"{HttpContext.Request.Scheme}://localhost:7165/Table/{table.Id.ToString()}";
+
            await _tableRepository.Create(table);
 
                 QRCodeHelper.GetQRCode(table.Link,20,Color.Black,Color.White,QRCodeGenerator.ECCLevel.M);
@@ -85,7 +87,11 @@ namespace Restraunt.WebAPI.Controllers
         [HttpDelete]
         public async Task<ActionResult<List<Table>>> DeleteTable(Guid Id)
         {
-          await _tableRepository.Delete(Id);
+            var entity = await _tableRepository.Get(Id);
+            if (entity == null)
+                return BadRequest("Id not found");
+
+            await _tableRepository.Delete(Id);
             return Ok(await _tableRepository.Select());
 
         }
